// play/chess-tourney-frontend/src/api/model/gameStatus.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GameStatus = {
  Waiting: "Waiting",
  InProgress: "InProgress",
  FinishedWin: "FinishedWin",
  FinishedDraw: "FinishedDraw",
  FinishedSemiDraw: "FinishedSemiDraw",
} as const;