// play/chess-tourney-frontend/src/api/model/move.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */
import type { Position } from "./position";

export interface Move {
  from: Position;
  to: Position;
  move_uci: string;
}