// play/chess-tourney-frontend/src/api/model/playerUpdate.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */
import type { PlayerUpdateWhite } from "./playerUpdateWhite";
import type { PlayerUpdateGrey } from "./playerUpdateGrey";
import type { PlayerUpdateBlack } from "./playerUpdateBlack";

export interface PlayerUpdate {
  white: PlayerUpdateWhite;
  grey: PlayerUpdateGrey;
  black: PlayerUpdateBlack;
}